Fix bug in argument passing
Because the Fiber block does not use splat:

```
@fiber = Fiber.new do |args|
...
end
```

only one argument was being passed from the `#run` method in the call to
`@fiber.resume(*args)`. This fixes the issue by removing the splat from
the call to `resume`. This could also be fixed by adding a splat to the
block parameter. I chose this method under the assumption that it will
be faster by skipping the parameterization step when it's not necessary,
but adding the splat to the block parameter might be more consistent
with the rest of the code.