Don't use `self.logger` for logging.
The `@logger` should only be used for carrying logger state between parent
and child tasks. But using it directly for logging is likely to produce
inconsistent output because the methods of the task might not execute in the
scope of the fiber itself. In the future, we hope it would be possible to
remove this variable completely and instead have a more direct relationship
between parent and child fiber locals.