Because `Node#stop` can ignore transient tasks, introduce `#terminate`.
Directly stopping a task still works as expected, but transient children are
ignored. The root node (reactor) does need an interface to stop all tasks,
so we introduce `Node#terminate` which recursively stops all children,
including transients.

We also make the `later` parameter to `#stop` explicit. There are several
ways in which a task can be stopped, and deferring this process to the next
iteration of the run loop can be necessary if the task cannot be stopped
immediately - i.e. if it's currently resumed.